refactor(types): rename API types for clarity and extract to separate file                - Adjust naming of CharacterSearchResponse and CharacterDataResponse to clearly reflect their purposes                - Move all API type definitions into src/store/types.ts for better organization
